requirements.txt
moviepy
azure-cognitiveservices-speech
python-dotenv
pydub